+++ libAudio/loadAON.cpp
@@ -61,8 +61,8 @@ bool modAON_t::isAON(const int32_t fd) noexcept
 	std::array<char, 42> aonMagic2;
 	return
 		fd != -1 &&
-		read(fd, aonMagic1.data(), aonMagic1.size()) == aonMagic1.size() &&
-		read(fd, aonMagic2.data(), aonMagic2.size()) == aonMagic2.size() &&
+		static_cast<size_t>(read(fd, aonMagic1.data(), aonMagic1.size())) == aonMagic1.size() &&
+		static_cast<size_t>(read(fd, aonMagic2.data(), aonMagic2.size())) == aonMagic2.size() &&
 		lseek(fd, 0, SEEK_SET) == 0 &&
 		std::equal(libAudio::aon::magic1.begin(), libAudio::aon::magic1.end(), aonMagic1.cbegin()) &&
 		(aonMagic1[3] == '4' || aonMagic1[3] == '8') &&
