+++ libAudio/loadFLAC.cpp
@@ -352,7 +352,7 @@ bool flac_t::isFLAC(const int32_t fd) noexcept
 {
 	std::array<char, 4> flacMagic;
 	if (fd == -1 ||
-		read(fd, flacMagic.data(), flacMagic.size()) != flacMagic.size() ||
+		static_cast<size_t>(read(fd, flacMagic.data(), flacMagic.size())) != flacMagic.size() ||
 		lseek(fd, 0, SEEK_SET) != 0)
 		return false;
 	if (flacMagic == libAudio::flac::oggMagic)
