+++ libAudio/loadIT.cpp
@@ -50,7 +50,7 @@ bool modIT_t::isIT(const int32_t fd) noexcept
 	std::array<char, 4> itMagic;
 	return
 		fd != -1 &&
-		read(fd, itMagic.data(), itMagic.size()) == itMagic.size() &&
+		static_cast<size_t>(read(fd, itMagic.data(), itMagic.size())) == itMagic.size() &&
 		lseek(fd, 0, SEEK_SET) == 0 &&
 		itMagic == libAudio::it::magic;
 }
