+++ libAudio/loadWAV.cpp
@@ -364,9 +364,9 @@ bool wav_t::isWAV(const int32_t fd) noexcept
 	std::array<char, 4> waveMagic;
 	return
 		fd != -1 &&
-		read(fd, riffMagic.data(), riffMagic.size()) == riffMagic.size() &&
+		static_cast<size_t>(read(fd, riffMagic.data(), riffMagic.size())) == riffMagic.size() &&
 		lseek(fd, 4, SEEK_CUR) == 8 &&
-		read(fd, waveMagic.data(), waveMagic.size()) == waveMagic.size() &&
+		static_cast<size_t>(read(fd, waveMagic.data(), waveMagic.size())) == waveMagic.size() &&
 		lseek(fd, 0, SEEK_SET) == 0 &&
 		riffMagic == libAudio::wave::riffMagic &&
 		waveMagic == libAudio::wave::waveMagic;
